Wed, Mar 2, 2022, 12:33 AM +00:00

+++ readability/readability.c
@@ -27,7 +27,7 @@ int main(void)
         }
     }
 
-    float L = (float)(letters / words);
+    float L = (float)letters / (float)words;
     float S = (float)((sentences / words) * 100);
     int index = round((0.0588 * L) - (0.296 * S) - 15.8);
